Rename NR microphysics plugin names
Add multiple simulation
