+++ appletree/plugins/__init__.py
@@ -4,8 +4,8 @@ from .common import *
 from . import er_microphys
 from .er_microphys import *
 
-from . import nr_microphys
-from .nr_microphys import *
+from . import nr_lyqy
+from .nr_lyqy import *
 
 from . import detector
 from .detector import *
